refactor: ContentType 을 common 패키지로 이동
